requirements.txt
google-cloud-dialogflow-cx==1.41.0
google-api-core==2.17.1
google-auth==2.29.0
protobuf==4.25.3
pathlib==1.0.1